refactor : 피드백 반영 1. purchaseAmount 원시값 포장 2. PurchaseDetail 팩터리 메서드 추가
