+++ src/test/kotlin/lotto/domain/LottoTicketTest.kt
@@ -119,7 +119,7 @@ class LottoTicketTest {
     fun `같은 숫자와 콤마 구분자로 된 문자열로 생성된 로또 티켓은 동일하다`() {
         val numbers = "1,2,3,4,5,6"
         val lottoTicket = LottoTicket.makeLottoTicket(numbers)
-        val number2 =  "1,2,3,4,5,6"
+        val number2 = "1,2,3,4,5,6"
         val lottoTicket2 = LottoTicket.makeLottoTicket(number2)
         lottoTicket shouldBeEqual lottoTicket2
     }
